+++ Sourcery/CodeGenerated/Typed.generated.swift
@@ -7,6 +7,7 @@ extension AssociatedValue {
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
     var actualTypeName: TypeName? { return typeName.actualTypeName ?? typeName }
     var isTuple: Bool { return typeName.isTuple }
+    var isClosure: Bool { return typeName.isClosure }
 }
 extension MethodParameter {
     var isOptional: Bool { return typeName.isOptional }
@@ -14,6 +15,7 @@ extension MethodParameter {
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
     var actualTypeName: TypeName? { return typeName.actualTypeName ?? typeName }
     var isTuple: Bool { return typeName.isTuple }
+    var isClosure: Bool { return typeName.isClosure }
 }
 extension TupleElement {
     var isOptional: Bool { return typeName.isOptional }
@@ -21,6 +23,7 @@ extension TupleElement {
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
     var actualTypeName: TypeName? { return typeName.actualTypeName ?? typeName }
     var isTuple: Bool { return typeName.isTuple }
+    var isClosure: Bool { return typeName.isClosure }
 }
 extension Typealias {
     var isOptional: Bool { return typeName.isOptional }
@@ -28,6 +31,7 @@ extension Typealias {
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
     var actualTypeName: TypeName? { return typeName.actualTypeName ?? typeName }
     var isTuple: Bool { return typeName.isTuple }
+    var isClosure: Bool { return typeName.isClosure }
 }
 extension Variable {
     var isOptional: Bool { return typeName.isOptional }
@@ -35,4 +39,5 @@ extension Variable {
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
     var actualTypeName: TypeName? { return typeName.actualTypeName ?? typeName }
     var isTuple: Bool { return typeName.isTuple }
+    var isClosure: Bool { return typeName.isClosure }
 }