+++ SourceryTests/Models/TypedSpec.generated.swift
@@ -25,6 +25,11 @@ class TypedSpec: QuickSpec {
                 expect(sut.value(forKeyPath: "isTuple") as? Bool).to(equal(true))
             }
 
+            it("can report closure type via KVC") {
+                let sut = AssociatedValue(typeName: TypeName("(Int) -> (Int)"))
+                expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
+            }
+
             it("can report actual type name via KVC") {
                 let sut = AssociatedValue(typeName: TypeName("Alias"))
                 expect(sut.value(forKeyPath: "actualTypeName") as? TypeName).to(equal(TypeName("Alias")))
@@ -49,6 +54,11 @@ class TypedSpec: QuickSpec {
                 sut.typeName.tuple = tuple
 
                 expect(sut.value(forKeyPath: "isTuple") as? Bool).to(equal(true))
+            }
+
+            it("can report closure type via KVC") {
+                let sut = MethodParameter(typeName: TypeName("(Int) -> (Int)"))
+                expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
             }
 
             it("can report actual type name via KVC") {
@@ -77,6 +87,11 @@ class TypedSpec: QuickSpec {
                 expect(sut.value(forKeyPath: "isTuple") as? Bool).to(equal(true))
             }
 
+            it("can report closure type via KVC") {
+                let sut = TupleElement(typeName: TypeName("(Int) -> (Int)"))
+                expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
+            }
+
             it("can report actual type name via KVC") {
                 let sut = AssociatedValue(typeName: TypeName("Alias"))
                 expect(sut.value(forKeyPath: "actualTypeName") as? TypeName).to(equal(TypeName("Alias")))
@@ -103,6 +118,11 @@ class TypedSpec: QuickSpec {
                 expect(sut.value(forKeyPath: "isTuple") as? Bool).to(equal(true))
             }
 
+            it("can report closure type via KVC") {
+                let sut = Typealias(typeName: TypeName("(Int) -> (Int)"))
+                expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
+            }
+
             it("can report actual type name via KVC") {
                 let sut = AssociatedValue(typeName: TypeName("Alias"))
                 expect(sut.value(forKeyPath: "actualTypeName") as? TypeName).to(equal(TypeName("Alias")))
@@ -127,6 +147,11 @@ class TypedSpec: QuickSpec {
                 sut.typeName.tuple = tuple
 
                 expect(sut.value(forKeyPath: "isTuple") as? Bool).to(equal(true))
+            }
+
+            it("can report closure type via KVC") {
+                let sut = Variable(typeName: TypeName("(Int) -> (Int)"))
+                expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
             }
 
             it("can report actual type name via KVC") {